refactor: app navigation 화면 책임 정리
Refs #69
