+++ app/src/main/java/com/team/yeogibeoryeo/navigation/AppRoute.kt
@@ -3,9 +3,6 @@ package com.team.yeogibeoryeo.navigation
 import kotlinx.serialization.Serializable
 
 @Serializable
-data object HomeRoute
-
-@Serializable
 data object MapRoute
 
 @Serializable
