--- app/src/main/java/com/team/yeogibeoryeo/navigation/HomeScreen.kt
@@ -1,163 +1,0 @@
-package com.team.yeogibeoryeo.navigation
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.LocationOn
-import androidx.compose.material.icons.outlined.Recycling
-import androidx.compose.material.icons.outlined.Search
-import androidx.compose.material.icons.outlined.Today
-import androidx.compose.material3.Button
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun HomeScreen(
-    onItemSearch: (String?) -> Unit,
-    onMapClick: () -> Unit,
-    onRegionalGuideClick: () -> Unit,
-    onFavoritesClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var query by remember { mutableStateOf("") }
-
-    LazyColumn(
-        modifier = modifier.fillMaxSize(),
-        contentPadding = PaddingValues(20.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        item {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text(
-                    text = "여기 버려",
-                    style = MaterialTheme.typography.headlineMedium,
-                    fontWeight = FontWeight.Bold,
-                )
-                Text(
-                    text = "품목 검색, 수거 장소, 지역별 배출 정보를 한 곳에서 확인하세요.",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-        }
-
-        item {
-            OutlinedTextField(
-                value = query,
-                onValueChange = { query = it },
-                modifier = Modifier.fillMaxWidth(),
-                leadingIcon = {
-                    Icon(
-                        imageVector = Icons.Outlined.Recycling,
-                        contentDescription = null,
-                    )
-                },
-                trailingIcon = {
-                    Button(onClick = { onItemSearch(query.takeIf { it.isNotBlank() }) }) {
-                        Icon(
-                            imageVector = Icons.Outlined.Search,
-                            contentDescription = null,
-                        )
-                    }
-                },
-                placeholder = { Text("버릴 품목을 검색해보세요") },
-                singleLine = true,
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-                keyboardActions = KeyboardActions(
-                    onSearch = { onItemSearch(query.takeIf { it.isNotBlank() }) },
-                ),
-            )
-        }
-
-        item {
-            HomeActionCard(
-                title = "수거 장소 찾기",
-                description = "동네명, 장소명, 주소로 가까운 수거함을 찾아요.",
-                icon = { Icon(Icons.Outlined.LocationOn, contentDescription = null) },
-                onClick = onMapClick,
-            )
-        }
-
-        item {
-            HomeActionCard(
-                title = "지역별 배출 가이드",
-                description = "우리 동네 배출 요일과 방법을 확인해요.",
-                icon = { Icon(Icons.Outlined.Today, contentDescription = null) },
-                onClick = onRegionalGuideClick,
-            )
-        }
-
-        item {
-            HomeActionCard(
-                title = "즐겨찾기",
-                description = "저장 기능은 후속 작업에서 연결할 예정입니다.",
-                icon = { Icon(Icons.Outlined.Search, contentDescription = null) },
-                onClick = onFavoritesClick,
-            )
-        }
-    }
-}
-
-@Composable
-private fun HomeActionCard(
-    title: String,
-    description: String,
-    icon: @Composable () -> Unit,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Card(
-        onClick = onClick,
-        modifier = modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surface,
-        ),
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(18.dp),
-            horizontalArrangement = Arrangement.spacedBy(16.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            icon()
-            Column(
-                modifier = Modifier.weight(1f),
-                verticalArrangement = Arrangement.spacedBy(4.dp),
-            ) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = description,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-        }
-    }
-}
